Authentication

+++ API-MasterClass/dev-camper-api/route-controller/course-controller.js
@@ -50,6 +50,7 @@ module.exports.getSingleCourses = asyncHandler(async (req, res, next) => {
 module.exports.createNewCourse = asyncHandler(async (req, res, next) => {
     const bootcampId = req.params.bootcampId;
     req.body.bootcamp = bootcampId;
+    req.body.user = req.user.id;
 
     const bootcamp = await Bootcamp.findById(bootcampId);
     if (!bootcamp) {
@@ -57,6 +58,14 @@ module.exports.createNewCourse = asyncHandler(async (req, res, next) => {
             new ErrorResponse(
                 `Bootcamp not found with id of ${req.params.id}`,
                 404
+            )
+        );
+    }
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(
+            new ErrorResponse(
+                `User ${req.user.id} is not authorized to add a course to ${bootcamp._id}`,
+                401
             )
         );
     }
@@ -81,7 +90,15 @@ module.exports.updateCourse = asyncHandler(async (req, res, next) => {
             )
         );
     }
-
+    //Make sure user is course owner
+    if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(
+            new ErrorResponse(
+                `User ${req.user.id} is not authorized to update the course course ${course._id}`,
+                401
+            )
+        );
+    }
     const updatedCourse = await Course.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -110,7 +127,15 @@ module.exports.deleteCourse = asyncHandler(async (req, res, next) => {
             )
         );
     }
-
+    //Make sure user is course owner
+    if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(
+            new ErrorResponse(
+                `User ${req.user.id} is not authorized to delete the course course ${course._id}`,
+                401
+            )
+        );
+    }
     course.deleteOne();
 
     res.status(200).json({
